Added Interest screen
I created the interest.py file that creates the interest screen. Now after logging in or skipping the login, this should now lead to the interest screen that has the capability to follow and unfollow interest bubbles. Very basic stuff but something to get a visual on the placeholders. I also corrected this screen to be centered.