Added create_bubble.py, feed.py, and profile.py
I created the a feed interface with a right navigation sidebar. It is very barebones but it'll have to do for now. I want to start adding comments to make it more readable. The profile, create bubble, and home placeholders buttons can be clicked on but the forums itself are just basic placeholders for now.